ReachingDefinitionsAnalysis: extract data model
300+ lines not related to the logic of the analysis could be moved into
their own file.
