+++ angr/analyses/reaching_definitions/__init__.py
@@ -1,4 +1,4 @@
-
-from .reaching_definitions import ReachingDefinitionAnalysis, LiveDefinitions
+from .reaching_definitions import ReachingDefinitionAnalysis
+from .live_definitions import LiveDefinitions
 from .def_use import DefUseAnalysis, DefUseState
 from .constants import OP_AFTER, OP_BEFORE
